requirements.txt
# Core dependencies
pydantic>=2.0.0
pyyaml>=6.0.0
python-dotenv>=1.0.0

# HTTP client
requests>=2.31.0
tenacity>=8.2.0
msgpack>=1.0.0  # Required by hyperliquid-python-sdk (explicit for clarity)

# Data processing
pandas>=2.0.0
pandas-ta>=0.3.14b0
numpy>=1.24.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-asyncio>=0.21.0

# Code quality
pylint>=3.0.0
mypy>=1.7.0
black>=23.12.0
isort>=5.13.0

# Logging
coloredlogs>=15.0.0

# Database (Phase 2+)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL driver
alembic>=1.13.0  # Database migrations

# Trading execution (Phase 3)
hyperliquid-python-sdk>=0.20.0  # Official HyperLiquid SDK
eth-account>=0.10.0  # Ethereum account (used by SDK)
web3>=6.11.0  # Web3 utilities
eth-utils>=2.3.0  # Ethereum utilities

# Automation (Phase 4)
APScheduler==3.10.4  # Job scheduling
click==8.1.7  # CLI framework
python-daemon==3.0.1  # Daemon process
tabulate==0.9.0  # Table formatting for CLI
psutil==5.9.6  # System metrics

# AI/LLM Integration (Phase 2)
openai>=1.0.0  # OpenAI-compatible API client (for DeepSeek and other LLMs)

# Future dependencies (Phase 5+)
# fastapi>=0.104.0  # For web API (Phase 5)
# uvicorn>=0.24.0  # For web server (Phase 5)
